Revert "Migrate to Authjs/NextAuth v5"
