+++ package.json
@@ -18,7 +18,7 @@
     "classnames": "^2.5.1",
     "geist": "^1.3.1",
     "next": "^14.2.13",
-    "next-auth": "^5.0.0-beta.22",
+    "next-auth": "^4.24.8",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-intersection-observer": "^9.13.1",
