+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,3 +1,6 @@
-import { handlers } from '@/auth';
+import NextAuth from 'next-auth';
+import authOptions from '@/lib/auth/auth-options';
 
-export const { GET, POST } = handlers;
+const handler = NextAuth(authOptions);
+
+export { handler as GET, handler as POST };
