+++ src/components/navbar.tsx
@@ -1,7 +1,7 @@
-import Link from 'next/link';
-
-import { auth, signOut } from '@/auth';
+import { SignOutButton } from '@/components/auth/auth-buttons';
+import auth from '@/lib/auth/auth';
 import Button from '@/components/ui/button';
+import Link from 'next/link';
 
 const Navbar = async () => {
   const session = await auth();
@@ -14,17 +14,7 @@ const Navbar = async () => {
         </Link>
       </Button>
       <nav className='flex gap-4'>
-        {session && (
-          <form
-            className='flex flex-col gap-4'
-            action={async () => {
-              'use server';
-              await signOut();
-            }}
-          >
-            <Button type='submit'>Signout</Button>
-          </form>
-        )}
+        {session && <SignOutButton>Sign out</SignOutButton>}
       </nav>
     </header>
   );
